+++ documentation/changelog.rst
@@ -11,6 +11,7 @@ v0.27.1 | July XX, 2025
 Bugfixes
 -----------
 * Flex-context price sensors can have any shared unit [see `PR #1628 <https://github.com/FlexMeasures/flexmeasures/pull/1628>`_]
+* Fix a few UI grievances in asset creation, esp. child assets [see `PR #1629 <https://github.com/FlexMeasures/flexmeasures/pull/1629>`_]
 
 
 v0.27.0 | July 20, 2025